+++ utils/storage.py
@@ -184,7 +184,7 @@ class Storage:
         return profile
 
     async def get_tag_profile(self, ctx, user: discord.Member):
-        query = f'SELECT steam,xbox,psn,nintendo,minecraft,origin,blizzard,bethesda FROM gamertags WHERE uid={user.id};'
+        query = f'SELECT steam,xbox,psn,nintendo,minecraft,origin,blizzard,bethesda,gog FROM gamertags WHERE uid={user.id};'
         r = self.execute_query(query)
         if not r or len(r) == 0:
             await self.create_profile(ctx, user)